Reduce proof dependencies on specific rules (like mergerule) and other high level concepts
