+++ key.core/src/test/java/de/uka/ilkd/key/proof/reference/TestReferenceSearcher.java
@@ -76,7 +76,8 @@ class TestReferenceSearcher {
         p.pruneProof(foundReference);
         p.closeGoal(p.getOpenGoal(foundReference));
         assertTrue(p.closed());
-        foundReference.proof().copyCachedGoals(p2, null, null);
+        Proof proof = foundReference.proof();
+        CopyReferenceResolver.copyCachedGoals(proof, p2, null, null);
         assertTrue(p.closed());
         GeneralSettings.noPruningClosed = true;
         p.dispose();
